+++ build.gradle
@@ -26,10 +26,15 @@ dependencies {
 	implementation 'org.springframework.boot:spring-boot-starter-data-jpa'
 	implementation 'org.springframework.boot:spring-boot-starter-web'
 	implementation 'org.springframework.boot:spring-boot-starter-websocket'
+
+
+
 	implementation 'mysql:mysql-connector-java:8.0.32'
     implementation 'org.jetbrains:annotations:23.0.0'
+    implementation 'top.jfunc.json:Json-Gson:1.0'
+	implementation 'com.google.code.gson:gson:2.10.1'
 
-    compileOnly 'org.projectlombok:lombok'
+	compileOnly 'org.projectlombok:lombok'
 	developmentOnly 'org.springframework.boot:spring-boot-devtools'
 	runtimeOnly 'com.h2database:h2'
 	annotationProcessor 'org.projectlombok:lombok'
