+++ src/main/java/com/example/demo/chat/dto/ChatHistoryDto.java
@@ -1,0 +1,26 @@
+package com.example.demo.chat.dto;
+
+
+import jakarta.persistence.EntityListeners;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+import org.springframework.format.annotation.DateTimeFormat;
+
+import java.time.LocalDateTime;
+
+@Builder
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+public class ChatHistoryDto {
+
+    String userId;
+    String userName;
+    String message;
+
+}
