+++ src/main/java/com/example/demo/chat/entity/ChatHistory.java
@@ -1,0 +1,32 @@
+package com.example.demo.chat.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.CreationTimestamp;
+
+import java.time.LocalDate;
+
+
+@Getter
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table(name = "chat_history")
+public class ChatHistory {
+
+    @Id
+    Long chatHistoryIdx;
+    String senderId;
+    String senderName;
+    String message;
+    @CreationTimestamp
+    LocalDate sendedDate;
+
+}
