+++ src/main/java/com/example/demo/chat/repository/ChatHistoryRepository.java
@@ -1,0 +1,10 @@
+package com.example.demo.chat.repository;
+
+import com.example.demo.chat.entity.ChatHistory;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface ChatHistoryRepository extends JpaRepository<ChatHistory, String> {
+
+}
