+++ src/main/java/com/example/demo/config/WebSocketConfig.java
@@ -1,13 +1,30 @@
 package com.example.demo.config;
 
+
+import com.example.demo.websocket.handler.SocketHandler;
 import lombok.RequiredArgsConstructor;
+import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.web.socket.WebSocketHandler;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
+import org.springframework.web.socket.config.annotation.WebSocketConfigurer;
+import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry;
 
 @Configuration
-@RequiredArgsConstructor
 @EnableWebSocket
-public class WebSocketConfig {
+@RequiredArgsConstructor
+public class WebSocketConfig implements WebSocketConfigurer {
 
-    
+
+    @Override
+    public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
+        registry.addHandler(socketHandler(), "/chat/{userId}")
+                .setAllowedOriginPatterns("*")
+                .withSockJS();
+    }
+
+    @Bean
+    public WebSocketHandler socketHandler() {
+        return new SocketHandler();
+    }
 }
