--- src/main/java/com/example/demo/controller/ChatHistoryController.java
@@ -1,4 +1,0 @@
-package com.example.demo.controller;
-
-public class ChatHistoryController {
-}
