--- src/main/java/com/example/demo/controller/CustomExceptionController.java
@@ -1,21 +1,0 @@
-package com.example.demo.controller;
-
-
-import com.example.demo.domain.CustomException;
-import com.example.demo.domain.dto.CustomExceptionResponse;
-import jakarta.servlet.http.HttpServletRequest;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.ErrorResponse;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-@RestControllerAdvice
-public class CustomExceptionController extends ResponseEntityExceptionHandler {
-
-    @ExceptionHandler(CustomException.class)
-    public ResponseEntity<CustomExceptionResponse> handleCustomException (CustomException customException) {
-        return CustomExceptionResponse.toResponseEntity(customException.getCustomErrorCode());
-    }
-
-}
