--- src/main/java/com/example/demo/domain/ChatHistory.java
@@ -1,28 +1,0 @@
-package com.example.demo.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.util.Date;
-
-@Getter
-@Entity
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table(name = "chat_history")
-public class ChatHistory {
-
-    @Id
-    String senderId;
-    String senderName;
-    String message;
-    Date sendedDate;
-
-}
