--- src/main/java/com/example/demo/domain/CustomException.java
@@ -1,13 +1,0 @@
-package com.example.demo.domain;
-
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException{
-
-    private final CustomErrorCode customErrorCode;
-
-}
