--- src/main/java/com/example/demo/domain/User.java
@@ -1,27 +1,0 @@
-package com.example.demo.domain;
-
-import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-import lombok.AccessLevel;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@Entity
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Table (name = "user")
-public class User {
-
-    @Id
-    String userId;
-    String userPassword;
-    String userName;
-    String userEmail;
-
-}
