--- src/main/java/com/example/demo/domain/UserStateCode.java
@@ -1,10 +1,0 @@
-package com.example.demo.domain;
-
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-@Getter
-@RequiredArgsConstructor
-public enum UserStateCode {
-    ON_LINE, OFF_LINE;
-}
