--- src/main/java/com/example/demo/domain/dto/OnlinedUserListDto.java
@@ -1,4 +1,0 @@
-package com.example.demo.domain.dto;
-
-public class OnlinedUserListDto {
-}
