--- src/main/java/com/example/demo/repository/ChatHistoryRepository.java
@@ -1,10 +1,0 @@
-package com.example.demo.repository;
-
-import com.example.demo.domain.ChatHistory;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface ChatHistoryRepository extends JpaRepository<ChatHistory, String> {
-
-}
