+++ src/main/java/com/example/demo/user/dto/UserListDto.java
@@ -1,0 +1,20 @@
+package com.example.demo.user.dto;
+
+import jakarta.persistence.EntityListeners;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.jpa.domain.support.AuditingEntityListener;
+
+@Builder
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+@EntityListeners(AuditingEntityListener.class)
+public class UserListDto {
+
+    String userId;
+    String userName;
+    boolean isOnline;
+}
