+++ src/main/java/com/example/demo/user/entity/User.java
@@ -1,0 +1,28 @@
+package com.example.demo.user.entity;
+
+import jakarta.persistence.Entity;
+import jakarta.persistence.Id;
+import jakarta.persistence.Table;
+import lombok.AccessLevel;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Entity
+@Builder
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Table (name = "user")
+public class User {
+
+    @Id
+    Long userIdx;
+    String userId;
+    String userPassword;
+    String userName;
+    String userEmail;
+
+}
